+++ redis/commands/timeseries/__init__.py
@@ -1,5 +1,5 @@
 import redis
-from redis.client import bool_ok
+from redis._parsers.helpers import bool_ok
 
 from ..helpers import parse_to_list
 from .commands import (
